+++ models/staging/synthea/stg_synthea__medications.sql
@@ -20,7 +20,7 @@ WITH cte_medications_lower AS (
         , encounter AS encounter_id
         , code AS medication_code
         , description AS medication_description
-        , {{ dbt.cast("base_cost", api.Column.translate_type("decimal")) }} AS medication_base_cost
+        , {{ dbt.cast("base_cost", api.Column.translate_type("float")) }} AS medication_base_cost
         , payer_coverage AS medication_payer_coverage
         , dispenses
         , totalcost AS medication_total_cost
