+++ models/staging/synthea/stg_synthea__procedures.sql
@@ -19,7 +19,7 @@ WITH cte_procedures_lower AS (
         , encounter AS encounter_id
         , code AS procedure_code
         , description AS procedure_description
-        , {{ dbt.cast("base_cost", api.Column.translate_type("decimal")) }} AS procedure_base_cost
+        , {{ dbt.cast("base_cost", api.Column.translate_type("float")) }} AS procedure_base_cost
         , reasoncode AS procedure_reason_code
         , reasondescription AS procedure_reason_description
     FROM cte_procedures_lower
